backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
alembic>=1.14,<2
httpx>=0.28,<1
python-dotenv>=1.0,<2
PyJWT>=2.8,<3
